+++ src/parser/statements.js
@@ -61,6 +61,12 @@ export const statementParserMethods = {
 				return this.parseFor();
 			case T.SWITCH:
 				return this.parseSwitch();
+			case T.GO:
+				this.err("goroutines are not supported in GoFront");
+				break;
+			case T.SELECT:
+				this.err("select statement is not supported in GoFront (no channels)");
+				break;
 			case T.LBRACE:
 				return this.parseBlock();
 			case T.IDENT: